Moved non-Chart plots into geom module
